Minor tweaks

+++ Breeze/NSManagedObject+Deleting.swift
@@ -11,7 +11,8 @@ import CoreData
 
 extension NSManagedObject {
     public func deleteInContextOfType(type: BreezeContextType) {
-        let contextSelf = inContextOfType(type) as! NSManagedObject
-        BreezeStore.contextForType(type).deleteObject(contextSelf)
+        if let contextSelf = inContextOfType(type) {
+            BreezeStore.contextForType(type).deleteObject(contextSelf)
+        }
     }
 }
